[FEATURE] add component to display uploaded job confirmations
The new JobConfirmationData dir includes the JobConfirmationDataCard
as the container for the views of the uploaded data. When there is
only one upload the data is display in an form wiht disabled input
fields. When more than one file is uploded an acordion with the
last uploaded file expanded is displayed.

Closes #1